[Fix] app.vueのtemplateの配置を変更
