+++ app/frontend/components/pages/error/ServerError.vue
@@ -24,11 +24,12 @@
 
 <script>
 import { mapActions } from 'vuex';
+
 export default {
   methods: {
-    ...mapActions('status', ['fetchResponseStatus']),
+    ...mapActions('responseState', ['fetchResponseState']),
     comeBack() {
-      this.fetchResponseStatus(null);
+      this.fetchResponseState({ status: '', state: '' });
       this.$router.go(-1);
     },
   },
