+++ app/frontend/store/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import users from './modules/users';
 import snackbars from './modules/snackbars';
-import status from './modules/status';
+import responseState from './modules/responseState';
 
 Vue.use(Vuex);
 
@@ -11,7 +11,7 @@ export default new Vuex.Store({
   strict: process.env.NODE_ENV !== 'production',
   modules: {
     users,
+    responseState,
     snackbars,
-    status,
   },
 });
